backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
scikit-learn==1.8.0
numpy>=2.0.0
pandas>=2.0.0
pydantic[email]>=2.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
psycopg2-binary>=2.9.10
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=8.3.3
pytest-asyncio>=0.24.0
httpx>=0.27.2
slowapi>=0.1.9
prometheus-client>=0.21.0
redis>=5.2.0
